feat(cli): add synchronous query agent and enhance CLI
- Introduce sql_synthesizer.sync_query_agent.QueryAgent for synchronous natural language to SQL translation with schema discovery and caching.
- Refactor sql_synthesizer.query_agent to use the new synchronous QueryAgent and improve CLI interface with argument parsing, error handling, and interactive mode.
- Add support for quantum-inspired query optimization toggle and related CLI options.
- Update entry points in pyproject.toml and setup.py to point to the new CLI main.
- Modify imports in sql_synthesizer/__init__.py and webapp.py to use the synchronous QueryAgent.
- Add TERRAGON_AUTONOMOUS_IMPLEMENTATION_REPORT.md documenting autonomous SDLC execution and production readiness.
- Improve logging configuration and cache backend initialization with fallback.
- Enhance CLI with options for schema cache TTL, query cache TTL, output CSV, SQL-only mode, and detailed error reporting.
- Add graceful handling of common exceptions and keyboard interrupts in CLI.
- Update tests import to reflect new module structure.

This change provides a robust synchronous query agent implementation and a more user-friendly, feature-rich CLI for the SQL Synthesizer project.
