+++ pyproject.toml
@@ -61,7 +61,7 @@ docs = [
 ]
 
 [project.scripts]
-query-agent = "query_agent:main"
+query-agent = "sql_synthesizer.query_agent:main"
 query-agent-web = "sql_synthesizer.webapp:main"
 
 # Development convenience scripts (use with: python -m <script>)
